LPS-56198 Moves EditExpandoAction, no logic changes
